fix(desktop): clear local mode push checks
